ENH: Add template for multimodal package (#333)
Added a folder and Python namespace for multimodal work, plus a simple PR workflow
Added an empty template folder "new_project_template" from which we can start future projects quicker
