+++ .github/workflows/build-test-pr.yml
@@ -249,7 +249,7 @@ jobs:
       - name: Install dependencies
         run: |
           cd hi-ml-azure
-          pip install ${{ steps.download_himlazure.outputs.package_filename }} 
+          pip install ${{ steps.download_himlazure.outputs.package_filename }}
           cd ../hi-ml
           pip install ${{ steps.download_himl.outputs.package_filename }}
 
@@ -264,7 +264,7 @@ jobs:
           branch_prefix="refs/heads/"
           full_branch_name=$GITHUB_REF
           branch_name_without_prefix=${full_branch_name#$branch_prefix}
-          python hi-ml/run_pytest.py --mark=gpu --cluster=pr-gpu --conda_env=${{ matrix.folder }}/environment.yml --folder=${{ matrix.folder }} --experiment="$branch_name_without_prefix"
+          python hi-ml-azure/run_pytest.py --mark=gpu --cluster=pr-gpu --conda_env=${{ matrix.folder }}/environment.yml --folder=${{ matrix.folder }} --experiment="$branch_name_without_prefix"
 
       - name: Upload coverage artifacts
         uses: ./.github/actions/upload_coverage_artifacts
