+++ hi-ml-azure/run_pytest.py
@@ -8,13 +8,22 @@ from _pytest.main import ExitCode
 
 # Add hi-ml packages to sys.path so that AML can find them if we are using the runner directly from the git repo
 himl_root = Path(__file__).resolve().parent.parent
-folders_to_add = [himl_root / "hi-ml" / "src", himl_root / "hi-ml-azure" / "src"]
-for folder in folders_to_add:
+
+
+def add_to_sys_path(folder: Path) -> None:
     folder_str = str(folder)
     if folder.is_dir() and folder_str not in sys.path:
         sys.path.insert(0, str(folder))
 
+
+folders_to_add = [himl_root / "hi-ml" / "src", himl_root / "hi-ml-azure" / "src"]
+for folder in folders_to_add:
+    add_to_sys_path(folder)
+
 from health_azure import submit_to_azure_if_needed  # noqa: E402
+from health_azure.himl import OUTPUT_FOLDER  # noqa: E402
+from health_azure.logging import logging_to_stdout  # noqa: E402
+from health_azure.paths import git_repo_root_folder  # noqa: E402
 from health_azure.utils import (  # noqa: E402
     WORKSPACE_CONFIG_JSON,
     check_config_json,
@@ -22,8 +31,6 @@ from health_azure.utils import (  # noqa: E402
     is_running_in_azure_ml,
     parse_arguments,
 )
-from health_ml.utils.common_utils import DEFAULT_AML_UPLOAD_DIR, logging_to_stdout  # noqa: E402
-from health_ml.utils.fixed_paths import repository_root_directory  # noqa: E402
 
 PYTEST_RESULTS_FILE = "pytest_results.xml"
 
@@ -46,6 +53,10 @@ class RunPytestConfig(param.Parameterized):
         default="30m", doc="The maximum runtime that is allowed for this job in AzureML. This is given as a floating"
         "point number with a string suffix s, m, h, d for seconds, minutes, hours, day. Examples: '3.5h', '2d'"
     )
+    add_to_sys_path: str = param.String(
+        default="",
+        doc="A folder name that should be added to sys.path. The folder name should be relative to repository root."
+    )
 
 
 def run_pytest(folder_to_test: str, pytest_mark: str) -> None:
@@ -57,7 +68,7 @@ def run_pytest(folder_to_test: str, pytest_mark: str) -> None:
     :param pytest_mark: The PyTest mark to use for filtering out the tests to run.
     :param folder_to_test: The folder with tests that should be run.
     """
-    results_file = Path(DEFAULT_AML_UPLOAD_DIR) / PYTEST_RESULTS_FILE
+    results_file = Path(OUTPUT_FOLDER) / PYTEST_RESULTS_FILE
     pytest_args = [folder_to_test, f"--junitxml={str(results_file)}"]
 
     if pytest_mark:
@@ -80,6 +91,8 @@ if __name__ == "__main__":
     )
     parser_results = parse_arguments(parser, fail_on_unknown_args=True)
     config = RunPytestConfig(**parser_results.args)
+    if config.add_to_sys_path:
+        add_to_sys_path(himl_root / config.add_to_sys_path)
     logging_to_stdout()
     submit_to_azureml = config.cluster != ""
     if submit_to_azureml and not is_running_in_azure_ml():
@@ -91,7 +104,7 @@ if __name__ == "__main__":
                 compute_cluster_name=config.cluster,
                 submit_to_azureml=submit_to_azureml,
                 wait_for_completion=True,
-                snapshot_root_directory=repository_root_directory(),
+                snapshot_root_directory=git_repo_root_folder(),
                 conda_environment_file=config.conda_env,
                 experiment_name=config.experiment,
                 max_run_duration=config.max_run_duration
