+++ hi-ml-azure/testazure/testazure/test_azure_util.py
@@ -1362,7 +1362,6 @@ from health_azure.utils import download_files_from_run_id""",
     render_and_run_test_script(tmp_path / "foo", RunTarget.AZUREML, extra_options, extra_args=[], expected_pass=True)
 
 
-@pytest.mark.skip(reason="We have an open PR with a fix for this already")
 def test_replace_directory(tmp_path: Path) -> None:
     extra_options = {
         "imports": """
@@ -1373,17 +1372,17 @@ from health_azure.utils import replace_directory
 """,
 
         "body": """
-output_dir = Path("outputs/test_outputs")
-output_dir.mkdir(parents=True, exist_ok=True)
-file_name = "hello.txt"
-(output_dir / file_name).write_text("Hello World!")
-assert (output_dir / file_name).exists()
-new_output_dir = output_dir.parent / "more_test_outputs"
+    output_dir = Path("outputs/test_outputs")
+    output_dir.mkdir(parents=True, exist_ok=True)
+    file_name = "hello.txt"
+    (output_dir / file_name).write_text("Hello World!")
+    assert (output_dir / file_name).exists()
+    new_output_dir = output_dir.parent / "more_test_outputs"
 
-replace_directory(output_dir, new_output_dir)
+    replace_directory(output_dir, new_output_dir)
 
-assert not output_dir.exists()
-assert (new_output_dir / file_name).exists()
+    assert not output_dir.exists()
+    assert (new_output_dir / file_name).exists()
 """
     }
