+++ himl-projects.code-workspace
@@ -10,6 +10,9 @@
       "path": "hi-ml-histopathology"
     },
     {
+      "path": "multimodal"
+    },
+    {
       "path": ".",
       "name": "ROOT",
     },
@@ -42,6 +45,7 @@
       "hi-ml/": true,
       "hi-ml-azure/": true,
       "hi-ml-histopathology/": true,
+      "multimodal/": true,
     },
     // The settings below are shared across all projects.
     // When making changes here, please ensure they are kept in sync
