+++ multimodal/.flake8
@@ -1,0 +1,4 @@
+[flake8]
+max-line-length = 120
+max-complexity = 25
+ignore = E731
