+++ multimodal/.mypy.ini
@@ -1,0 +1,11 @@
+[mypy]
+python_version=3.7
+scripts_are_modules=True
+namespace_packages=True
+show_traceback=True
+ignore_missing_imports=True
+follow_imports=normal
+follow_imports_for_stubs=True
+disallow_untyped_calls=False
+disallow_untyped_defs=True
+strict_optional=True
