+++ multimodal/README.md
@@ -1,0 +1,11 @@
+# Multimodal
+
+## Developer setup
+
+Make file commands:
+
+* `make pip_local` to install multimodal package in editable mode. This must happen before running tests.
+* `make build` to build the package
+* `make mypy` to run `mypy`
+* `make check` to run `flake8`, `mypy` and `pyright`
+* `make clean` to clean up all temporary files and folders
