+++ multimodal/environment.yml
@@ -1,0 +1,24 @@
+name: multimodal
+channels:
+  - defaults
+  - pytorch
+dependencies:
+  - pip=20.1.1
+  - python=3.7.3
+  - pip:
+    # This file should contain pinned versions of all packages in requirements_run.txt
+    # and requirements_test.txt.
+    # It is not possible to reference the requirements files here with "-r" because AzureML's
+    # environment creation can't handle those.
+    # Run requirements: presently empty
+    # Test requirements
+    - black==22.1.0
+    - coverage==6.3.2
+    - flake8==4.0.1
+    - mypy==0.931
+    - pylint==2.12.2
+    - pycobertura==2.0.1
+    - pytest==6.2.2
+    - pytest-cov==2.11.1
+    - pytest-rerunfailures==10.2
+    - pytest-timeout==2.0.1
