+++ multimodal/package_description.md
@@ -1,0 +1,28 @@
+# Microsoft Health Intelligence Multimodal Toolbox
+
+## Overview
+
+This toolbox provides models for working with multi-modal health data.
+
+## Installation
+
+You can install the latest version from `pypi` via
+
+```shell
+pip install hi-ml-multimodal
+```
+
+## Documentation
+
+The detailed package documentation, with examples and API reference, is on
+[readthedocs](https://hi-ml.readthedocs.io/en/latest/).
+
+## Getting started
+
+Examples that illustrate the use of the toolbox can be found on
+[readthedocs](https://hi-ml.readthedocs.io/en/latest/).
+
+## Links
+
+* Github [https://github.com/microsoft/hi-ml](https://github.com/microsoft/hi-ml)
+* Project InnerEye [http://aka.ms/InnerEye](http://aka.ms/InnerEye)
