+++ multimodal/pytest.ini
@@ -1,0 +1,4 @@
+[pytest]
+testpaths = test_multimodal
+log_cli = True
+log_cli_level = DEBUG
