+++ multimodal/requirements_build.txt
@@ -1,0 +1,9 @@
+docutils==0.16
+myst-parser==0.15.2
+setuptools==59.5.0
+sphinx==4.1.2
+sphinx-autodoc-typehints==1.12.0
+sphinx-automodapi==0.13
+sphinx-rtd-theme==1.0.0
+twine==3.3.0
+wheel==0.36.2
