+++ new_project_template/.vscode/launch.json
@@ -1,0 +1,20 @@
+{
+    // Use IntelliSense to learn about possible attributes.
+    // Hover to view descriptions of existing attributes.
+    // For more information, visit: https://go.microsoft.com/fwlink/?linkid=830387
+    "version": "0.2.0",
+    "configurations": [
+        {
+            "name": "Setup package",
+            "type": "python",
+            "request": "launch",
+            "program": "${workspaceFolder}/setup.py",
+            "cwd": "${workspaceFolder}/",
+            "args": [
+                "sdist",
+                "bdist_wheel"
+            ],
+            "console": "integratedTerminal"
+        },
+    ]
+}
