+++ new_project_template/health_newproject/dummy.py
@@ -1,0 +1,8 @@
+#  -------------------------------------------------------------------------------------------
+#  Copyright (c) Microsoft Corporation. All rights reserved.
+#  Licensed under the MIT License (MIT). See LICENSE in the repo root for license information.
+#  -------------------------------------------------------------------------------------------
+
+
+def dummy() -> int:
+    return 1
