+++ new_project_template/pyrightconfig.json
@@ -1,0 +1,10 @@
+{
+  "include": [
+    "health_newproject",
+    "test_newproject",
+  ],
+  "useLibraryCodeForTypes": false,
+  "reportMissingImports": true,
+  "reportMissingTypeStubs": false,
+  "reportPrivateImportUsage": false,
+}
