+++ new_project_template/pytest.ini
@@ -1,0 +1,7 @@
+[pytest]
+testpaths = test_newproject
+log_cli = True
+log_cli_level = DEBUG
+addopts = --strict-markers
+markers =
+    gpu: Tests that require at least 1 GPU to be present
