+++ new_project_template/requirements_run.txt
@@ -1,0 +1,1 @@
+param>=1.12
